+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 # Change Log
 All notable changes to QETLAB will be documented in this file.
 
-## [0.8] - Changes since 2015-04-13
+## Changes since 2015-04-13
+### Added
+- BCSGameLB: Computes a lower bound on the quantum value of a binary constraint system (BCS) game.
+- BCSGameValue: Computes the maximum value of a binary constraint system (BCS) game. In classical and no-signalling settings, the value computed is exact, but the quantum value is just an upper bound.
+- NonlocalGameLB: Computes a lower bound on the quantum value of a two-player non-local game.
+- helpers/bcs_to_nonlocal: Converts a description of a binary constraint system (BCS) game into a form that can be presented as a general non-local game.
+
+### Changed
+- kpNorm: Can now be used as the objective function or as a constraint in a CVX optimization problem, regardless of k and p (only certain special values of k and p were supported previously).
+- SchattenNorm: Can now be used as the objective function or as a constraint in a CVX optimization problem, regardless of p (only p = 1, p = 2, and p = Inf were supported previously).
+
+## [0.8] - 2015-04-13
 ### Added
 - BellInequalityMaxQubits: Approximates the optimal value of a Bell inequality in qubit (i.e., 2-dimensional quantum) settings.
 - NonlocalGameValue: Computes the maximum value of a nonlocal game in a classical, quantum, or no-signalling setting.